pyproject.toml
[project]
name = "tourag2"
version = "0.1.0"
description = "Agentic RAG system for tours"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.34.0",
    "python-dotenv>=1.1.0",
    "smolagents>=1.18.0",
]
